marketing-mix-model/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
xgboost>=1.4.2
matplotlib>=3.4.2
seaborn>=0.11.1
jupyter>=1.0.0
pyyaml>=5.4.1
statsmodels>=0.12.2
pytest>=6.2.5